Split AnyLayerTangentVector to separate file
